requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.0

# Config
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==24.1.0
